refactor: Split private and public context app
