+++ src/components/AppPublic.jsx
@@ -1,0 +1,66 @@
+import React, { useEffect, useMemo, useState } from 'react'
+import Spinner from 'cozy-ui/transpiled/react/Spinner'
+import { useClient } from 'cozy-client'
+
+import { Editor, Unshared } from 'components/notes'
+import { fetchIfIsNoteReadOnly } from 'lib/utils'
+import { getReturnUrl, getSharedDocument } from 'lib/utils'
+import { AppLayout } from './AppLayout'
+
+const AppPublic = () => {
+  const client = useClient()
+  const [sharedDocumentId, setSharedDocumentId] = useState(null)
+  const [readOnly, setReadOnly] = useState(false)
+  const returnUrl = useMemo(() => getReturnUrl(), [])
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const searchParams = new URLSearchParams(window.location.search)
+        if (searchParams.has('id')) {
+          // public route = /public/?sharecode=xxxx&id=xxxxx
+          // The id of the note is necessary because the sharecode
+          // may be about a folder and not a specific note.
+          const id = searchParams.get('id')
+          const readOnly = await fetchIfIsNoteReadOnly(client, id)
+          setReadOnly(readOnly)
+          setSharedDocumentId(id)
+        } else {
+          // public route = /public/?sharecode=xxxxx
+          // There is no id. It should be a sharecode
+          // dedicated to a unique note. We look into
+          // permissions and try to open the first file.
+          const { id, readOnly } = await getSharedDocument(client)
+          setReadOnly(readOnly)
+          setSharedDocumentId(id)
+        }
+      } catch {
+        setSharedDocumentId(false)
+      }
+    }
+    fetchData()
+  }, [client])
+  if (sharedDocumentId) {
+    return (
+      <Editor
+        readOnly={readOnly}
+        noteId={sharedDocumentId}
+        returnUrl={returnUrl}
+      />
+    )
+  } else if (sharedDocumentId !== null) {
+    return <Unshared />
+  } else {
+    return <Spinner size="xxlarge" middle />
+  }
+}
+
+const AppPublicWrapper = () => {
+  return (
+    <AppLayout>
+      <AppPublic />
+    </AppLayout>
+  )
+}
+
+export default AppPublicWrapper
