+++ src/targets/public/index.jsx
@@ -11,7 +11,7 @@ import { AppProviders } from 'components/AppProviders'
 import { initApp } from 'lib/initApp'
 
 const renderApp = function(appLocale, client) {
-  const App = require('components/app').default
+  const App = require('components/AppPublic').default
   const isPublic = true
 
   render(
